# docs/swagger.yaml
basePath: /v1
definitions:
...
  users.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      last_name:
        type: string
      no_of_games:
        type: integer
      phone:
        type: string
      profile_picture_url:
        type: string
      skill_level:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact:
    email: Ongchen10sherpa@gmail.com
    name: fullstacksherpa
    url: https://www.fullstacksherpa.tech/
  description: API for Khel, a complete sport application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Khel API
paths:
...
  /users:
    put:
      consumes:
      - application/json
      description: Update user information such as first name, last name, skill level,
        and phone number
      parameters:
      - description: 'Request body containing fields to update: first_name, last_name,
          skill_level, phone'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "204":
          description: User info updated successfully
          schema:
            type: string
        "400":
          description: Bad request, update values can't be nil
          schema: {}
        "404":
          description: User not found
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Update user information
      tags:
      - users
  /users/{userID}/follow:
    put:
      consumes:
      - application/json
      description: Follows a user by ID
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: User followed
          schema:
            type: string
        "400":
          description: User payload missing
          schema: {}
        "404":
          description: User not found
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Follows a user
      tags:
      - users
  /users/{userID}/unfollow:
    put:
      consumes:
      - application/json
      description: Unfollow a user by ID
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: User unfollowed
          schema:
            type: string
        "400":
          description: User payload missing
          schema: {}
        "404":
          description: User not found
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Unfollow a user
      tags:
      - users
  /users/activate/{token}:
    put:
      consumes:
      - application/json
      description: Activate a user account using an activation token provided in the
        URL
      parameters:
      - description: Activation token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: User activated
          schema:
            type: string
        "404":
          description: User not found
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Activate user account
      tags:
      - authentication
...
  /users/me:
    delete:
      description: Deletes the logged-in user's account and Cloudinary profile photo
      produces:
      - application/json
      responses:
        "204":
          description: User deleted
          schema:
            type: string
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Delete current user account
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Retrieve the authenticated user’s profile information
      produces:
      - application/json
      responses:
        "200":
          description: Current user data
          schema:
            $ref: '#/definitions/users.User'
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get current user profile
      tags:
      - users
  /users/profile-picture:
    post:
      consumes:
      - multipart/form-data
      description: Uploads a user's profile picture and saves the URL in the database
      parameters:
      - description: Profile picture file size limit is 2MB
        in: formData
        name: profile_picture
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 'Profile picture uploaded successfully: <URL>'
          schema:
            type: string
        "400":
          description: Unable to parse form or retrieve file
          schema: {}
        "500":
          description: Failed to upload image to Cloudinary or save URL in database
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Upload profile picture
      tags:
      - users
    put:
      consumes:
      - multipart/form-data
      description: Updates a user's profile picture, saves the new URL in the database,
        and deletes the old one from Cloudinary
      parameters:
      - description: 'Profile picture file (max size: 2MB)'
        in: formData
        name: profile_picture
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 'Profile picture updated successfully: <URL>'
          schema:
            type: string
        "400":
          description: Unable to parse form or retrieve file
          schema: {}
        "500":
          description: Failed to upload image to Cloudinary, update database, or delete
            old image
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Update profile picture
      tags:
      - users
...
  /users/update-profile:
    patch:
      consumes:
      - multipart/form-data
      description: Update any combination of first name, last name, phone, skill level,
        and/or profile picture in one call.
      parameters:
      - description: First name
        in: formData
        name: first_name
        type: string
      - description: Last name
        in: formData
        name: last_name
        type: string
      - description: Phone number (10 digits)
        in: formData
        name: phone
        type: string
      - description: Skill level
        enum:
        - beginner
        - intermediate
        - advanced
        in: formData
        name: skill_level
        type: string
      - description: JPEG or PNG image (max 5 MB)
        in: formData
        name: profile_picture
        type: file
      produces:
      - application/json
      responses:
        "204":
          description: Profile updated successfully
          schema:
            type: string
        "400":
          description: Bad request (e.g. parse error, invalid field)
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Edit current user’s profile
      tags:
      - users
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"